Create Tonya Gauna
I know this is late and if you don't want to accept it then I understand.  I didn't stop until I could get it to work.  The only thing I could not get to work was the moving to the caretta directory.  I am not sure what I did wrong.